AITPROJECT/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
redis
python-kucoin
python-dotenv
pandas
joblib
scikit-learn==1.5.2
asyncio
sse-starlette
aiofiles
broadcaster
websockets
sqlalchemy